third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/dedicated-worker.https-expected.txt

This is a testharness.js-based test.
PASS COEP: none worker in COEP: none frame
PASS COEP: none worker in COEP: require-corp frame
FAIL COEP: require-corp worker in COEP: none frame assert_equals: expected "FAILED" but got "LOADED"
PASS COEP: require-corp worker in COEP: require-corp frame
PASS COEP: none module worker in COEP: none frame
PASS COEP: none module worker in COEP: require-corp frame
FAIL COEP: require-corp module worker in COEP: none frame assert_equals: expected "FAILED" but got "LOADED"
PASS COEP: require-corp module worker in COEP: require-corp frame
Harness: the test ran to completion.
